backend/requirements.txt
Flask
requests
python-dotenv
psycopg2-binary
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
